+++ src/include/LSP/ClientConfiguration.hpp
@@ -57,6 +57,7 @@ struct ClientInlayHintsConfiguration
     bool variableTypes = false;
     bool parameterTypes = false;
     bool functionReturnTypes = false;
+    bool hideHintsForErrorTypes = false;
     size_t typeHintMaxLength = 50;
     /// Whether type inlay hints should be made insertable
     bool makeInsertable = true;
@@ -64,7 +65,8 @@ struct ClientInlayHintsConfiguration
     inline bool operator==(const ClientInlayHintsConfiguration& rhs) const
     {
         return this->parameterNames == rhs.parameterNames && this->variableTypes == rhs.variableTypes && this->parameterTypes == rhs.parameterTypes &&
-               this->functionReturnTypes == rhs.functionReturnTypes && this->typeHintMaxLength == rhs.typeHintMaxLength;
+               this->functionReturnTypes == rhs.functionReturnTypes && this->hideHintsForErrorTypes == rhs.hideHintsForErrorTypes &&
+               this->typeHintMaxLength == rhs.typeHintMaxLength;
     }
 
     inline bool operator!=(const ClientInlayHintsConfiguration& rhs) const
@@ -72,8 +74,8 @@ struct ClientInlayHintsConfiguration
         return !(*this == rhs);
     }
 };
-NLOHMANN_DEFINE_TYPE_NON_INTRUSIVE_WITH_DEFAULT(
-    ClientInlayHintsConfiguration, parameterNames, variableTypes, parameterTypes, functionReturnTypes, typeHintMaxLength, makeInsertable);
+NLOHMANN_DEFINE_TYPE_NON_INTRUSIVE_WITH_DEFAULT(ClientInlayHintsConfiguration, parameterNames, variableTypes, parameterTypes, functionReturnTypes,
+    hideHintsForErrorTypes, typeHintMaxLength, makeInsertable);
 
 struct ClientHoverConfiguration
 {